+++ libs/resiliency/tests/performance/replay/1d_stencil_checksum.cpp
@@ -72,7 +72,7 @@ public:
         {
             data_[k] = std::sin(2 * pi *
                 ((0.0 + subdomain_width * subdomain_index + k) /
-                    (subdomain_width * subdomains)));
+                    static_cast<double>(subdomain_width * subdomains)));
             checksum_ += data_[k];
         }
     }
@@ -336,7 +336,9 @@ int hpx_main(hpx::program_options::variables_map& vm)
     }
 
     std::cout << "Time elapsed: "
-              << (hpx::util::high_resolution_clock::now() - t) / 1e9
+              << static_cast<double>(
+                     hpx::util::high_resolution_clock::now() - t) /
+            1e9
               << std::endl;
     std::cout << "Errors occurred: " << counter << std::endl;
